boac-5924 - fix issue with is_user_in_peer_advising_department
